+++ apps/graph/graph/integral_graph_controller.cpp
@@ -41,7 +41,7 @@ Layout IntegralGraphController::createFunctionLayout(ExpiringPointer<Shared::Fun
   constexpr size_t bufferSize = SymbolAbstract::k_maxNameSize+5; // f(x)dx
   char buffer[bufferSize];
   const char * dx = "dx";
-  int numberOfChars = function->nameWithArgument(buffer, bufferSize-strlen(dx));
+  size_t numberOfChars = function->nameWithArgument(buffer, bufferSize-strlen(dx));
   assert(numberOfChars <= bufferSize);
   strlcpy(buffer+numberOfChars, dx, bufferSize-numberOfChars);
   return LayoutHelper::String(buffer, strlen(buffer), KDFont::SmallFont);
